.docker/ticker/requirements.txt
pymongo
python-dateutil